requirements.txt
asgiref==3.8.1
Django==4.2.20
django-nine==0.2.7
django-strawberry==0.1.2
exceptiongroup==1.2.2
graphql-core==3.2.6
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
pytest-django==4.10.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
strawberry-graphql==0.210.0
tomli==2.2.1
typing_extensions==4.12.2
